- silence gcc liberal enum value handling warnings
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code/trunk/gwyddion@11603 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ libgwyddion/gwyexpr.c
@@ -736,7 +736,7 @@ gwy_expr_scan_tokens(GwyExpr *expr,
 
     scanner = expr->scanner;
     while ((token = g_scanner_get_next_token(scanner))) {
-        switch (token) {
+        switch ((gint)token) {
             case G_TOKEN_LEFT_PAREN:
             case G_TOKEN_RIGHT_PAREN:
             case G_TOKEN_COMMA:
